+++ tests/templates/kuttl/smoke/test_zookeeper.py
@@ -3,6 +3,7 @@ import argparse
 import requests
 import time
 import sys
+
 sys.tracebacklimit = 0
 
 
@@ -37,17 +38,29 @@ def check_ruok(hosts):
         url = host + ":8080/commands/" + cmd_ruok
         response = try_get(url).json()
 
-        if "command" in response and response["command"] == cmd_ruok \
-                and "error" in response and response["error"] is None:
+        if (
+            "command" in response
+            and response["command"] == cmd_ruok
+            and "error" in response
+            and response["error"] is None
+        ):
             continue
         else:
-            print("Error[" + cmd_ruok + "] for [" + url + "]: received " + str(
-                response) + " - expected {'command': 'ruok', 'error': None} ")
+            print(
+                "Error["
+                + cmd_ruok
+                + "] for ["
+                + url
+                + "]: received "
+                + str(response)
+                + " - expected {'command': 'ruok', 'error': None} "
+            )
             exit(-1)
 
 
 def check_monitoring(hosts):
     for host in hosts:
+        # test for the jmx exporter metrics
         url = host + ":9505"
         response = try_get(url)
 
@@ -57,16 +70,46 @@ def check_monitoring(hosts):
             print("Error for [" + url + "]: could not access monitoring")
             exit(-1)
 
+        # test for the native metrics
+        url = host + ":7000/metrics"
+        response = try_get(url)
 
-if __name__ == '__main__':
+        if response.ok:
+            # arbitrary metric was chosen to test if metrics are present in the response
+            if "quorum_size" in response.text:
+                continue
+            else:
+                print("Error for [" + url + "]: missing metrics")
+                exit(-1)
+            continue
+        else:
+            print("Error for [" + url + "]: could not access monitoring")
+            exit(-1)
+
+
+if __name__ == "__main__":
     all_args = argparse.ArgumentParser(description="Test ZooKeeper.")
-    all_args.add_argument("-n", "--namespace", help="The namespace to run in", required=True)
+    all_args.add_argument(
+        "-n", "--namespace", help="The namespace to run in", required=True
+    )
     args = vars(all_args.parse_args())
     namespace = args["namespace"]
 
-    host_primary_0 = "http://test-zk-server-primary-0.test-zk-server-primary." + namespace + ".svc.cluster.local"
-    host_primary_1 = "http://test-zk-server-primary-1.test-zk-server-primary." + namespace + ".svc.cluster.local"
-    host_secondary = "http://test-zk-server-secondary-0.test-zk-server-secondary." + namespace + ".svc.cluster.local"
+    host_primary_0 = (
+        "http://test-zk-server-primary-0.test-zk-server-primary."
+        + namespace
+        + ".svc.cluster.local"
+    )
+    host_primary_1 = (
+        "http://test-zk-server-primary-1.test-zk-server-primary."
+        + namespace
+        + ".svc.cluster.local"
+    )
+    host_secondary = (
+        "http://test-zk-server-secondary-0.test-zk-server-secondary."
+        + namespace
+        + ".svc.cluster.local"
+    )
 
     hosts = [host_primary_0, host_primary_1, host_secondary]